+++ src/dftbp/dftb/energytypes.F90
@@ -32,7 +32,7 @@ module dftbp_dftb_energytypes
     !> Spin energy
     real(dp) :: Espin = 0.0_dp
 
-    !> Range-separation energy
+    !> Hybrid xc-functional energy
     real(dp) :: Efock = 0.0_dp
 
     !> Spin orbit energy
@@ -68,7 +68,7 @@ module dftbp_dftb_energytypes
     !> Onsite correction energy
     real(dp) :: eOnSite = 0.0_dp
 
-    !> Onsite correction energy from range-separated functional
+    !> Onsite correction energy from hybrid xc-functional
     real(dp) :: EfockOnSite = 0.0_dp
 
     !> Halogen-X correction energy